# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/favorite_applications": {
            "get": {
                "tags": [
                    "executor"
                ],
                "summary": "Favorite applications",
                "description": "Get user's favorite applications",
                "operationId": "74170e18fdfb1e19df5aa971c395ec9d",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "genre": {
                                            "type": "string",
                                            "example": "Натюрморт"
                                        },
                                        "user": {
                                            "type": "string",
                                            "example": "Иван Иванов"
                                        },
                                        "user_avatar": {
                                            "type": "string",
                                            "example": "/storage/img/avatar.png"
                                        },
                                        "application_category": {
                                            "type": "string",
                                            "example": "Для всех"
                                        },
                                        "description": {
                                            "type": "text",
                                            "example": "Нарисовать натюрморт"
                                        },
                                        "payment": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "5000"
                                        },
                                        "writing_technique": {
                                            "type": "string",
                                            "example": "Акварель"
                                        },
                                        "deadline": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "20.04.2022"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "3 weeks ago"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/favorite_applications/{id}": {
            "post": {
                "tags": [
                    "executor"
                ],
                "summary": "Favorite applications",
                "description": "Add favorite applications",
                "operationId": "85d78ecc936421a5eb2d45ee098a211e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Application Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Добавлено в избранное"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    }
}